pyproject.toml
[tool.poetry]
name = "pentov1"
version = "0.1.0"
description = "Pentov1 using crewAI for security assessment"
authors = ["V Dundigalla <v.dundigalla6776@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<=3.13"
crewai = "^0.11.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
langchain-core = "^0.1.0"
langchain-mistralai = "^0.0.3"
mistralai = "^0.0.11"
python-dotenv = "^1.0.0"
zaproxy = "^0.3.2"
pydantic = "^2.5.2"
requests = "^2.28.0"
PyJWT = "^2.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pentov1 = "pentov1.main:run"
    